Improve doc and file structure
